requirements.txt
amqp==5.1.1
asgiref==3.6.0
astroid==2.13.5
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
black==23.1.0
boto3==1.34.64
botocore==1.34.64
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==3.0.4
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.32.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==39.0.2
daphne==3.0.2
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
dj-database-url==1.0.0
dj-email-url==1.0.6
dj-static==0.0.6
Django==4.1.2
django-cache-url==3.4.4
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-decorator-include==3.0
django-dotenv==1.4.2
django-filter==22.1
django-import-export==3.0.1
django-rest-swagger==2.2.0
django-smart-selects==1.6.0
django-storages==1.14.2
django-timezone-field==5.0
djangorestframework==3.14.0
drf-yasg==1.21.4
environs==9.5.0
et-xmlfile==1.1.0
flake8==6.0.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lazy-object-proxy==1.9.0
MarkupPy==1.14
MarkupSafe==2.1.2
marshmallow==3.19.0
mccabe==0.7.0
msgpack==1.0.5
mypy-extensions==1.0.0
oauthlib==3.2.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.1
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
prompt-toolkit==3.0.38
# psycopg2
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pylint==2.15.7
pyOpenSSL==23.0.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-magic==0.4.27
python-multipart==0.0.5
pytz==2022.7.1
redis==4.4.2
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
static3==0.7.0
tablib==3.3.0
tomli==2.0.1
tomlkit==0.11.6
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.12
validators==0.20.0
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.2.0
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.5.2
djangorestframework-camel-case
pendulum
Pillow==10.2.0
factory-boy==3.3.0
stripe==8.7.0
sphinx_rtd_theme==2.0.0
sphinx==7.2.6